Refactor to push code into its own files
